ItemRadio styled component migration
