+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **[BREAKING CHANGE]** `<Tabs>`: changed export for `TabsStatus`
 - **[BREAKING CHANGE]** `Button`: changed export for `Button.STATUS` to new `ButtonStatus`
 - **[BREAKING CHANGE]** `DatePicker`: changed export of `DatePicker.constants` to `DatePickerOrientation`
+- **[BREAKING CHANGE]** `<ItemRadio>`: changed export for `ItemRadioStatus`
 
 # v8.8.0 (24/07/2019)
