+++ src/index.tsx
@@ -15,7 +15,7 @@ export { default as ItemCheckbox } from 'itemCheckbox'
 export { default as ItemChoice } from 'itemChoice'
 export { default as ItemData } from 'itemData'
 export { default as ItemInfo } from 'itemInfo'
-export { default as ItemRadio } from 'itemRadio'
+export { default as ItemRadio, ItemRadioStatus } from 'itemRadio/ItemRadio'
 export { default as ItemRadioGroup } from 'itemRadioGroup'
 export { default as ItemsList } from 'itemsList'
 export { default as Itinerary } from 'itinerary'
